+++ src/embed/liveboard.ts
@@ -195,6 +195,19 @@ export class LiveboardEmbed extends V1Embed {
     };
 
     /**
+     * Triggers an event to the embedded app
+     * @param messageType The event type
+     * @param data The payload to send with the message
+     */
+    public trigger(messageType: HostEvent, data: any = {}): Promise<any> {
+        const dataWithVizId = data;
+        if (this.viewConfig.vizId) {
+            dataWithVizId.vizId = this.viewConfig.vizId;
+        }
+        return super.trigger(messageType, dataWithVizId);
+    }
+
+    /**
      * Render an embedded ThoughtSpot Liveboard or visualization
      * @param renderOptions An object specifying the Liveboard ID,
      * visualization ID and the runtime filters.