+++ src/embed/search.spec.ts
@@ -1,4 +1,5 @@
 import { SearchEmbed, HiddenActionItemByDefaultForSearchEmbed } from './search';
+import * as authInstance from '../auth';
 import { init } from '../index';
 import { Action, AuthType } from '../types';
 import {
@@ -20,7 +21,10 @@ const answerId = 'eca215d4-0d2c-4a55-90e3-d81ef6848ae0';
 const thoughtSpotHost = 'tshost';
 const defaultParams = `hostAppUrl=local-host&viewPortHeight=768&viewPortWidth=1024&sdkVersion=${version}`;
 const hideBydefault = `&hideAction=${fixedEncodeURI(
-    JSON.stringify(HiddenActionItemByDefaultForSearchEmbed),
+    JSON.stringify([
+        Action.ReportError,
+        ...HiddenActionItemByDefaultForSearchEmbed,
+    ]),
 )}`;
 const defaultParamsWithHiddenActions = defaultParams + hideBydefault;
 const prefixParams = '&isSearchEmbed=true';
@@ -30,11 +34,15 @@ beforeAll(() => {
         thoughtSpotHost,
         authType: AuthType.None,
     });
+    spyOn(window, 'alert');
 });
 
 describe('Search embed tests', () => {
     beforeEach(() => {
         document.body.innerHTML = getDocumentBody();
+        jest.spyOn(authInstance, 'getReleaseVersion').mockReturnValue(
+            '7.4.0.sw',
+        );
     });
 
     test('should render', async () => {
@@ -197,6 +205,7 @@ describe('Search embed tests', () => {
         searchEmbed.render();
         const hideActionUrl = fixedEncodeURI(
             JSON.stringify([
+                Action.ReportError,
                 ...hiddenActionsForSearch,
                 ...HiddenActionItemByDefaultForSearchEmbed,
             ]),
@@ -220,6 +229,7 @@ describe('Search embed tests', () => {
         searchEmbed.render();
         const hideActionUrl = fixedEncodeURI(
             JSON.stringify([
+                Action.ReportError,
                 ...hiddenActionsForSearch,
                 ...HiddenActionItemByDefaultForSearchEmbed,
             ]),