+++ src/react/index.spec.tsx
@@ -2,15 +2,17 @@ import React from 'react';
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
 import { cleanup, fireEvent, render, waitFor } from '@testing-library/react';
+import { Action, EmbedEvent, HostEvent } from '../types';
 import {
     executeAfterWait,
     getIFrameEl,
     getIFrameSrc,
     postMessageToParent,
+    mockMessageChannel,
 } from '../test/test-utils';
 import { SearchEmbed, AppEmbed, LiveboardEmbed, useEmbedRef } from './index';
 import { AuthType, init } from '../index';
-import { EmbedEvent, HostEvent } from '../types';
+
 import { version } from '../../package.json';
 
 const thoughtSpotHost = 'localhost';
@@ -20,6 +22,7 @@ beforeAll(() => {
         thoughtSpotHost,
         authType: AuthType.None,
     });
+    spyOn(window, 'alert');
 });
 
 describe('React Components', () => {
@@ -37,7 +40,7 @@ describe('React Components', () => {
                 ),
             ).toBe(true);
             expect(getIFrameSrc(container)).toBe(
-                `http://${thoughtSpotHost}/?hostAppUrl=local-host&viewPortHeight=768&viewPortWidth=1024&sdkVersion=${version}&hideAction=[%22editACopy%22,%22saveAsView%22,%22updateTSL%22,%22editTSL%22,%22onDeleteAnswer%22]&dataSourceMode=hide&useLastSelectedSources=false&isSearchEmbed=true#/embed/answer`,
+                `http://${thoughtSpotHost}/?hostAppUrl=local-host&viewPortHeight=768&viewPortWidth=1024&sdkVersion=${version}&hideAction=[%22${Action.ReportError}%22,%22editACopy%22,%22saveAsView%22,%22updateTSL%22,%22editTSL%22,%22onDeleteAnswer%22]&dataSourceMode=hide&useLastSelectedSources=false&isSearchEmbed=true#/embed/answer`,
             );
         });
 
@@ -73,6 +76,7 @@ describe('React Components', () => {
     describe('LiveboardEmbed', () => {
         //
         it('Should be able to trigger events on the embed using refs', async () => {
+            mockMessageChannel();
             const TestComponent = () => {
                 const embedRef = useEmbedRef();
                 const onLiveboardRendered = () => {
@@ -109,6 +113,7 @@ describe('React Components', () => {
                         data: ['viz1', 'viz2'],
                     },
                     `http://${thoughtSpotHost}`,
+                    expect.anything(),
                 );
             });
         });