+++ src/test/test-utils.ts
@@ -70,3 +70,19 @@ export const EVENT_WAIT_TIME = 1000;
 export function fixedEncodeURI(str: string) {
     return encodeURI(str).replace(/%5B/g, '[').replace(/%5D/g, ']');
 }
+
+/**
+ * MessageChannel is available in Node > 15.0.0. Since the current node environment's
+ * used for github actions is not above 14, we are mocking this for the current unit tests.
+ */
+export const messageChannelMock: any = {
+    port1: {},
+    port2: {},
+};
+export const mockMessageChannel = () => {
+    messageChannelMock.port1.close = jest.fn();
+    messageChannelMock.port2.onmessage = jest.fn();
+    window.MessageChannel = function MessageChannelMock() {
+        return messageChannelMock;
+    } as any;
+};