+++ src/utils/processTrigger.ts
@@ -1,3 +1,4 @@
+import { ERROR_MESSAGE } from '../errors';
 import { HostEvent } from '../types';
 
 /**
@@ -17,24 +18,47 @@ function postIframeMessage(
     iFrame: HTMLIFrameElement,
     message: { type: HostEvent; data: any },
     thoughtSpotHost: string,
+    channel?: MessageChannel,
 ) {
-    return iFrame.contentWindow.postMessage(message, thoughtSpotHost);
+    return iFrame.contentWindow.postMessage(message, thoughtSpotHost, [
+        channel?.port2,
+    ]);
 }
+
+const TRIGGER_TIMEOUT = 30000;
 
 export function processTrigger(
     iFrame: HTMLIFrameElement,
     messageType: HostEvent,
     thoughtSpotHost: string,
     data: any,
-) {
-    switch (messageType) {
-        case HostEvent.Reload:
-            return reload(iFrame);
-        default:
-            return postIframeMessage(
-                iFrame,
-                { type: messageType, data },
-                thoughtSpotHost,
-            );
-    }
+): Promise<any> {
+    return new Promise<any>((res, rej) => {
+        if (messageType === HostEvent.Reload) {
+            reload(iFrame);
+            return res(null);
+        }
+        const channel = new MessageChannel();
+        channel.port1.onmessage = ({ data: responseData }) => {
+            channel.port1.close();
+            if (responseData.error) {
+                rej(responseData.error);
+            } else {
+                res(responseData);
+            }
+        };
+
+        // Close the messageChannel and resolve the promise if timeout.
+        setTimeout(() => {
+            channel.port1.close();
+            res(new Error(ERROR_MESSAGE.TRIGGER_TIMED_OUT));
+        }, TRIGGER_TIMEOUT);
+
+        return postIframeMessage(
+            iFrame,
+            { type: messageType, data },
+            thoughtSpotHost,
+            channel,
+        );
+    });
 }